+++ packages/manager/src/components/AccessPanel/AccessPanel.tsx
@@ -78,7 +78,7 @@ interface Props {
   className?: string;
   small?: boolean;
   isOptional?: boolean;
-  hideHelperText?: boolean;
+  passwordHelperText?: string;
 }
 
 export interface UserSSHKeyObject {
@@ -111,7 +111,7 @@ class AccessPanel extends React.Component<CombinedProps> {
       className,
       small,
       isOptional,
-      hideHelperText,
+      passwordHelperText,
       requestKeys
     } = this.props;
 
@@ -141,7 +141,7 @@ class AccessPanel extends React.Component<CombinedProps> {
               placeholder={placeholder || 'Enter a password.'}
               onChange={this.handleChange}
               hideStrengthLabel={hideStrengthLabel}
-              hideHelperText={hideHelperText}
+              helperText={passwordHelperText}
             />
           </React.Suspense>
           {users && (