+++ packages/manager/src/features/linodes/LinodesDetail/LinodeRebuild/RebuildFromStackScript.test.tsx
@@ -39,8 +39,10 @@ const props: CombinedProps = {
   imagesError: {},
   imagesLastUpdated: 0,
   userSSHKeys: [],
+  disabled: false,
   closeSnackbar: jest.fn(),
   enqueueSnackbar: jest.fn(),
+  passwordHelperText: '',
   ...reactRouterProps
 };
 
@@ -68,7 +70,7 @@ describe('RebuildFromStackScript', () => {
       () => [
         getByText('A StackScript is required.'),
         getByText('An image is required.'),
-        getByText('Password cannot be blank.')
+        getByText('Password is required.')
       ],
       {}
     );